+++ custom_components/solarcharger/state_machine/state_charge.py
@@ -14,8 +14,7 @@ from ..const import (
     ENTITY_CHARGER_CHARGING_SENSOR,
     ENTITY_CHARGER_GET_CHARGE_CURRENT,
     NUMBER_CHARGER_MAX_SPEED,
-    NUMBER_POWER_MONITOR_DURATION,
-    SENSOR_CONSUMED_POWER,
+    SENSOR_CHARGER_ALLOCATED_POWER,
     ChargeStatus,
     RunState,
 )
@@ -50,122 +49,6 @@ class StateCharge(SolarChargeState):
 
     # ----------------------------------------------------------------------------
     # Subscriptions
-    # ----------------------------------------------------------------------------
-    async def _async_adjust_charge_current(
-        self, charger: Charger, chargeable: Chargeable, allocated_power: float
-    ) -> None:
-        """Adjust charge current."""
-
-        new_charge_current, old_charge_current = self._calc_current_change(
-            charger, chargeable, allocated_power, self.solarcharge.running_goal
-        )
-        if new_charge_current != old_charge_current:
-            _LOGGER.info(
-                "%s: Update current from %s to %s",
-                self.solarcharge.caller,
-                old_charge_current,
-                new_charge_current,
-            )
-            await self.solarcharge.async_set_charge_current(charger, new_charge_current)
-            self.solarcharge.charge_current_updatetime = utcnow().timestamp()
-
-        # No need to update status here because it is now done at the main loop.
-        # Power update here is not garanteed because HA will not send the same value
-        # if the second value is same as first.
-        # await self._async_update_ha(chargeable)
-
-    # ----------------------------------------------------------------------------
-    # 2025-11-02 09:01:48.009 INFO (MainThread) [custom_components.solarcharger.chargers.controller] tesla_custom_tesla23m3:
-    # entity_id=number.solarcharger_tesla_custom_tesla23m3_charger_allocated_power,
-    #
-    # old_state=<state number.solarcharger_tesla_custom_tesla23m3_charger_allocated_power=-500.0; min=-23000.0, max=23000.0, step=1.0, mode=box,
-    # unit_of_measurement=W, device_class=power, icon=mdi:flash, friendly_name=tesla_custom Tesla23m3 Allocated power @ 2025-11-02T20:00:32.962356+11:00>,
-    #
-    # new_state=<state number.solarcharger_tesla_custom_tesla23m3_charger_allocated_power=-200.0; min=-23000.0, max=23000.0, step=1.0, mode=box,
-    # unit_of_measurement=W, device_class=power, icon=mdi:flash, friendly_name=tesla_custom Tesla23m3 Allocated power @ 2025-11-02T20:01:48.008211+11:00>
-    async def _async_handle_allocated_power_update(
-        self, event: Event[EventStateChangedData]
-    ) -> None:
-        """Fetch and process state change event."""
-        data = event.data
-        entity_id = data["entity_id"]
-        old_state = data["old_state"]
-        new_state = data["new_state"]
-
-        if new_state is not None:
-            duration_since_last_change = (
-                new_state.last_changed_timestamp
-                - self.solarcharge.charge_current_updatetime
-            )
-
-            if old_state is not None:
-                _LOGGER.debug(
-                    "%s: entity_id=%s, old_state=%s, new_state=%s, duration_since_last_change=%s",
-                    self.solarcharge.caller,
-                    entity_id,
-                    old_state.state,
-                    new_state.state,
-                    duration_since_last_change,
-                )
-            else:
-                _LOGGER.debug(
-                    "%s: entity_id=%s, new_state=%s, duration_since_last_change=%s",
-                    self.solarcharge.caller,
-                    entity_id,
-                    new_state.state,
-                    duration_since_last_change,
-                )
-
-            # Make sure we don't change the charge current too often
-            if duration_since_last_change >= MIN_TIME_BETWEEN_UPDATE:
-                try:
-                    allocated_power = float(new_state.state)
-
-                    # Save allocated power to calculate moving average.
-                    if self.solarcharge.max_allocation_count > 0:
-                        if not self.solarcharge.is_monitor_available_power():
-                            self.solarcharge.power_allocations = []
-                        elif (
-                            len(self.solarcharge.power_allocations)
-                            >= self.solarcharge.max_allocation_count
-                        ):
-                            self.solarcharge.power_allocations.pop(0)
-
-                        assert self.solarcharge.entities.sensors is not None
-                        consumed_power = float(
-                            self.solarcharge.entities.sensors[
-                                SENSOR_CONSUMED_POWER
-                            ].state
-                        )
-                        self.solarcharge.power_allocations.append(
-                            allocated_power - consumed_power
-                        )
-
-                    # This callback is still running in paused state.
-                    # Only adjust charge current if we are still in charging state
-                    if self.solarcharge.machine_state.state == RunState.CHARGING:
-                        await self._async_adjust_charge_current(
-                            self.solarcharge.charger,
-                            self.solarcharge.chargeable,
-                            allocated_power,
-                        )
-
-                except Exception as e:
-                    _LOGGER.exception(
-                        "%s: Failed to adjust current for net power %s W: %s",
-                        self.solarcharge.caller,
-                        new_state.state,
-                        e,
-                    )
-
-    # ----------------------------------------------------------------------------
-    def _subscribe_allocated_power_update(self) -> None:
-        """Subscribe for allocated power update."""
-
-        self.solarcharge.tracker.track_allocated_power_update(
-            self._async_handle_allocated_power_update
-        )
-
     # ----------------------------------------------------------------------------
     async def async_handle_soc_update(
         self, event: Event[EventStateChangedData]
@@ -340,6 +223,30 @@ class StateCharge(SolarChargeState):
         return (new_charge_current, old_charge_current)
 
     # ----------------------------------------------------------------------------
+    async def _async_adjust_charge_current(
+        self, charger: Charger, chargeable: Chargeable, allocated_power: float
+    ) -> None:
+        """Adjust charge current."""
+
+        new_charge_current, old_charge_current = self._calc_current_change(
+            charger, chargeable, allocated_power, self.solarcharge.running_goal
+        )
+        if new_charge_current != old_charge_current:
+            _LOGGER.info(
+                "%s: Update current from %s to %s",
+                self.solarcharge.caller,
+                old_charge_current,
+                new_charge_current,
+            )
+            await self.solarcharge.async_set_charge_current(charger, new_charge_current)
+            self.solarcharge.charge_current_updatetime = utcnow().timestamp()
+
+        # No need to update status here because it is now done at the main loop.
+        # Power update here is not garanteed because HA will not send the same value
+        # if the second value is same as first.
+        # await self._async_update_ha(chargeable)
+
+    # ----------------------------------------------------------------------------
     def _log_charging_status(self, charger: Charger, msg: str) -> None:
         """Generate debug message only if required."""
 
@@ -412,35 +319,12 @@ class StateCharge(SolarChargeState):
                     await self.solarcharge.async_turn_off_calibrate_max_charge_speed_switch()
 
     # ----------------------------------------------------------------------------
-    def _set_max_allocation_count(self, wait_net_power_update: float) -> None:
-        """Initialize max allocation count."""
-
-        monitor_duration = self.solarcharge.option_get_entity_number_or_abort(
-            NUMBER_POWER_MONITOR_DURATION
-        )
-        monitor_duration_seconds = monitor_duration * 60
-
-        # Set to 0 to disable power monitoring by default.
-        self.solarcharge.max_allocation_count = 0
-
-        if monitor_duration > 0:
-            if monitor_duration_seconds > (2 * wait_net_power_update):
-                self.solarcharge.max_allocation_count = round(
-                    monitor_duration_seconds / wait_net_power_update
-                )
-            else:
-                _LOGGER.error(
-                    "%s: Power monitor duration (%s minutes) must be more than 2 times longer than net power update interval (%s seconds)",
-                    self.solarcharge.caller,
-                    monitor_duration,
-                    wait_net_power_update,
-                )
-
-    # ----------------------------------------------------------------------------
     async def _switch_on_charger_and_set_current(
         self, charger: Charger, chargeable: Chargeable
     ) -> None:
         """Switch on charger and set initial current."""
+
+        self.solarcharge.participate_in_real_power_allocation()
 
         self.solarcharge.starting_goal = self.solarcharge.running_goal
         self.solarcharge.scheduler.log_goal(
@@ -449,7 +333,7 @@ class StateCharge(SolarChargeState):
         await self.solarcharge.async_turn_charger_switch(charger, turn_on=True)
         await self.solarcharge.async_set_charge_current(charger, INITIAL_CHARGE_CURRENT)
         await self.solarcharge.async_update_ha(chargeable)
-        self._subscribe_allocated_power_update()
+        # self._subscribe_allocated_power_update()
         self._log_charging_status(charger, "Charger ON")
 
     # ----------------------------------------------------------------------------
@@ -461,13 +345,6 @@ class StateCharge(SolarChargeState):
         stats: ChargeStats,
     ) -> ContextData:
         """Loop to charge device."""
-
-        # Init power monitor duration
-        self.solarcharge.wait_net_power_update = (
-            self.solarcharge.get_wait_net_power_update()
-        )
-        self._set_max_allocation_count(self.solarcharge.wait_net_power_update)
-        self.solarcharge.power_allocations = []
 
         # Initialise counts before starting loop
         stats.loop_success_count = 0
@@ -496,6 +373,19 @@ class StateCharge(SolarChargeState):
                     await self._switch_on_charger_and_set_current(charger, chargeable)
                     done_switch_on_charger = True
 
+                # Adjust current.
+                assert self.solarcharge.entities.sensors is not None
+                allocated_power = float(
+                    self.solarcharge.entities.sensors[
+                        SENSOR_CHARGER_ALLOCATED_POWER
+                    ].state
+                )
+                await self._async_adjust_charge_current(
+                    charger,
+                    chargeable,
+                    allocated_power,
+                )
+
                 # Check if calibration is required during charge.
                 await self._async_calibrate_max_charge_speed_if_required(
                     charger, chargeable
@@ -521,7 +411,7 @@ class StateCharge(SolarChargeState):
             # Sleep before re-evaluating charging conditions.
             # Charging state must be "charging" for loop_count > 0.
             # Tesla BLE need 25 seconds here, ie. OPTION_WAIT_CHARGEE_UPDATE_HA = 25 seconds
-            await asyncio.sleep(self.solarcharge.wait_net_power_update)
+            await self.solarcharge.async_charger_sleep()
             stats.loop_total_count += 1
 
         return context