+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 ### Breaking change
 - Reminder: Please [delete then re-add the integration](https://github.com/flashg1/SolarCharger/wiki/Configuration#how-to-delete-and-re-add-the-solarcharger-integration) if upgrading from v0.5.4 or prior.
 ### Improvement
+- Separated net power update interval from charge current update interval.
 - Added missing entities for config debug logging during startup.
 
 ## v0.6.1 2026-04-26